Reverting to commit New Version
Reverting to before the Conversion, because I messed up doing it
properly, and didn't test along the way, and altered stuff half way
through - big mess. Starting back at the last working version, and
changing (then testing) each bit at a time.

+++ Last Prototype/Last Prototype/Sprite.h
@@ -7,19 +7,14 @@ A sprite is an EntityContainer that loops through an animation based on a cycle.
 class Sprite : public EntityContainer
 {
 public:
-	Sprite(int x, int y) : EntityContainer(x, y), cycle(0) {}
+	Sprite(float x, float y) : EntityContainer(x, y), cycle(0) {}
 	~Sprite(void) {}
-
-	// Set this sprite's skin, and render
 	void render(void) { set_skin(); EntityContainer::render(); }
 
 protected:
-	int cycle; // The index cycle through a cycling animation
-	int max_cycles; // The max cycle value, after which cycle resets to 0
+	int cycle;
+	int max_cycles;
 	
-	// Increment the cycle up to its max_cycle, then drop back to 0
 	virtual void IncCycle(void) { cycle = (cycle >= (max_cycles-1)) ? 0 : cycle+1; }
-
-	// Set the skin (clip) of this sprite
 	virtual void set_skin(void) = 0;
 };
